.github/requirements.txt
python-dotenv
requests
pyyaml
python-gitlab
pywinrm
paramiko